refactor email a tiny bit code so mailgun_send can be used by gather.
